requirements.txt
Flask==3.1.0
mysql-connector-python==9.2.0
bcrypt==4.3.0
python-dotenv==1.1.0
qrcode==8.2
Pillow==12.1.1
gunicorn==23.0.0
